requirements.txt
configparser
pandas
requests
psycopg2-binary